move key shortcut logic to a separate file to clean things up
